+++ parser/geo_annotation.go
@@ -1,21 +1,12 @@
 package parser
 
 import (
-	"context"
-	"fmt"
-	"log"
-	"reflect"
-	"runtime"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/m-lab/annotation-service/api"
-	v2 "github.com/m-lab/annotation-service/api/v2"
 	"github.com/m-lab/etl/annotation"
 	"github.com/m-lab/etl/web100"
-
-	"cloud.google.com/go/bigquery"
 
 	"github.com/m-lab/etl/metrics"
 	"github.com/m-lab/etl/schema"
@@ -159,138 +150,4 @@ func AddGeoDataPTHop(hop *schema.ParisTracerouteHop, timestamp time.Time) {
 		metrics.AnnotationErrorCount.With(prometheus.
 			Labels{"source": "PT Hop had no dest_ip!"}).Inc()
 	}
-}
-
-// AddGeoDataNDTConnSpec takes a connection spec and a timestamp and
-// annotates the connection spec with geo data associated with each IP
-// Address. It will either sucessfully add the geo data or fail
-// silently and make no changes.
-// Deprecated:  Use Annotatable interface and parser.Base instead.
-func AddGeoDataNDTConnSpec(spec schema.Web100ValueMap, timestamp time.Time) {
-	// Time the response
-	timerStart := time.Now()
-	defer func(tStart time.Time) {
-		metrics.AnnotationTimeSummary.
-			With(prometheus.Labels{"test_type": "NDT"}).
-			Observe(float64(time.Since(tStart).Nanoseconds()))
-	}(timerStart)
-
-	GetAndInsertTwoSidedGeoIntoNDTConnSpec(spec, timestamp)
-}
-
-// CopyStructToMap takes a POINTER to an arbitrary SIMPLE struct and copies
-// it's fields into a value map. It will also make fields entirely
-// lower case, for convienece when working with exported structs. Also,
-// NEVER pass in something that is not a pointer to a struct, as this
-// will cause a panic.
-func CopyStructToMap(sourceStruct interface{}, destinationMap map[string]bigquery.Value) {
-	structToCopy := reflect.ValueOf(sourceStruct).Elem()
-	typeOfStruct := structToCopy.Type()
-	for i := 0; i < typeOfStruct.NumField(); i++ {
-		f := structToCopy.Field(i)
-		v := f.Interface()
-		switch t := v.(type) {
-		case string:
-			// TODO - are these still needed?  Does the omitempty cover it?
-			if t == "" {
-				continue
-			}
-		case int64:
-			if t == 0 {
-				continue
-			}
-		}
-		jsonTag, ok := typeOfStruct.Field(i).Tag.Lookup("json")
-		name := strings.ToLower(typeOfStruct.Field(i).Name)
-		if ok {
-			tags := strings.Split(jsonTag, ",")
-			if len(tags) > 0 && tags[0] != "" {
-				name = tags[0]
-			}
-		}
-		destinationMap[strings.ToLower(name)] = v
-	}
-}
-
-// GetAndInsertTwoSidedGeoIntoNDTConnSpec takes a timestamp and an
-// NDT connection spec. It will either insert the data into the
-// connection spec or silently fail.
-// TODO - should make a large batch request for an entire insert buffer.
-// See sidestream implementation for example.
-func GetAndInsertTwoSidedGeoIntoNDTConnSpec(spec schema.Web100ValueMap, timestamp time.Time) {
-	// TODO: Make metrics for sok and cok failures. And double check metrics for cleanliness.
-	cip, cok := spec.GetString([]string{"client_ip"})
-	sip, sok := spec.GetString([]string{"server_ip"})
-	reqData := make([]string, 0, 2)
-	if cok {
-		cip, _ = web100.NormalizeIPv6(cip)
-		reqData = append(reqData, cip)
-	} else {
-		metrics.AnnotationWarningCount.With(prometheus.
-			Labels{"source": "Missing client side IP."}).Inc()
-	}
-	if sok {
-		sip, _ = web100.NormalizeIPv6(sip)
-		reqData = append(reqData, sip)
-	} else {
-		metrics.AnnotationWarningCount.With(prometheus.
-			Labels{"source": "Missing server side IP."}).Inc()
-	}
-	if cok || sok {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		deadline, _ := ctx.Deadline()
-		defer cancel()
-		resp, err := v2.GetAnnotations(ctx, annotation.BatchURL, timestamp, reqData)
-		if err != nil {
-			if err.Error() == "context canceled" {
-				// These are NOT timeouts, and the ctx.Err() is nil.
-				timeRemaining := deadline.Sub(time.Now())
-				log.Println("context canceled, time remaining =", timeRemaining, " ctx err:", ctx.Err())
-				_, file, line, _ := runtime.Caller(0)
-				metrics.AnnotationErrorCount.With(prometheus.Labels{"source": fmt.Sprintf("context canceled %s:%d", file, line)}).Inc()
-			} else {
-				// There are many error types returned here, so we log the error, but use the code location
-				// for the metric.
-				log.Println(err)
-				_, file, line, _ := runtime.Caller(0)
-				metrics.AnnotationErrorCount.With(prometheus.Labels{"source": fmt.Sprint(file, ":", line)}).Inc()
-			}
-			return
-		}
-
-		if cok {
-			if data, ok := resp.Annotations[cip]; ok && data.Geo != nil {
-				CopyStructToMap(data.Geo, spec.Get("client_geolocation"))
-				if data.Network != nil {
-					asn, err := data.Network.BestASN()
-					if err != nil {
-						log.Println(err)
-					} else {
-						spec.Get("client").Get("network")["asn"] = asn
-					}
-				}
-			} else {
-				metrics.AnnotationErrorCount.With(prometheus.
-					Labels{"source": "Couldn't get geo data for the client side."}).Inc()
-			}
-		}
-		if sok {
-			if data, ok := resp.Annotations[sip]; ok && data.Geo != nil {
-				CopyStructToMap(data.Geo, spec.Get("server_geolocation"))
-				if data.Network != nil {
-					asn, err := data.Network.BestASN()
-					if err != nil {
-						log.Println(err)
-					} else {
-						spec.Get("server").Get("network")["asn"] = asn
-					}
-				}
-			} else {
-				metrics.AnnotationErrorCount.With(prometheus.
-					Labels{"source": "Couldn't get geo data for the server side."}).Inc()
-			}
-
-		}
-	}
-
 }
