+++ parser/geo_annotation_test.go
@@ -5,15 +5,12 @@ package parser_test
 import (
 	"fmt"
 	"io/ioutil"
-	"log"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
 	"strings"
 	"testing"
 	"time"
-
-	"cloud.google.com/go/bigquery"
 
 	"github.com/go-test/deep"
 	"github.com/m-lab/annotation-service/api"
@@ -297,210 +294,6 @@ func TestAddGeoDataPTHop(t *testing.T) {
 		p.AddGeoDataPTHop(&test.hop, test.timestamp)
 		if !reflect.DeepEqual(test.hop, test.res) {
 			t.Errorf("Expected %v, got %v for test %s", test.res, test.hop, test.url)
-		}
-	}
-}
-
-func testTime() time.Time {
-	// Note: this timestamp corresponds to the "h3d0c0" in the result strings.
-	tst, _ := time.Parse(time.RFC3339, "2002-10-02T15:00:00Z")
-	return tst
-}
-
-func TestAddGeoDataNDTConnSpec(t *testing.T) {
-	var tests = []struct {
-		spec      schema.Web100ValueMap
-		timestamp time.Time
-		url       string
-		res       schema.Web100ValueMap
-	}{
-		{
-			spec: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				spec["client_ip"] = "127.0.0.1"
-				spec["server_ip"] = "1.0.0.127"
-				return spec
-			}(),
-			timestamp: testTime(),
-			url:       "/10583?",
-			res: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				spec["client_ip"] = "127.0.0.1"
-				spec["server_ip"] = "1.0.0.127"
-				geoc := spec.Get("client_geolocation")
-				geoc["country_code"] = "US"
-				geoc["country_code3"] = "USA"
-				geoc["country_name"] = "United States of America"
-				geoc["region"] = "NY"
-				geoc["city"] = "Scarsdale"
-				geoc["area_code"] = int64(10583)
-				geoc["postal_code"] = "10583"
-				geoc["latitude"] = float64(41.0051)
-				geoc["longitude"] = float64(73.7846)
-				geos := spec.Get("server_geolocation")
-				geos["country_code"] = "US"
-				geos["country_code3"] = "USA"
-				geos["country_name"] = "United States of America"
-				geos["region"] = "NY"
-				geos["city"] = "Scarsdale"
-				geos["area_code"] = int64(10584)
-				geos["postal_code"] = "10584"
-				geos["latitude"] = float64(41.0051)
-				geos["longitude"] = float64(73.7846)
-				return spec
-			}(),
-		},
-		{ // This test exercises the error path one missing IP
-			spec: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				spec["client_ip"] = "127.0.0.1"
-				return spec
-			}(),
-			timestamp: testTime(),
-			url:       "/10583?",
-			res: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				spec["client_ip"] = "127.0.0.1"
-				geoc := spec.Get("client_geolocation")
-				geoc["country_code"] = "US"
-				geoc["country_code3"] = "USA"
-				geoc["country_name"] = "United States of America"
-				geoc["region"] = "NY"
-				geoc["city"] = "Scarsdale"
-				geoc["area_code"] = int64(10583)
-				geoc["postal_code"] = "10583"
-				geoc["latitude"] = float64(41.0051)
-				geoc["longitude"] = float64(73.7846)
-				return spec
-			}(),
-		},
-		{ // This test exercises the error path for missing IP addresses.
-			spec: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				return spec
-			}(),
-			timestamp: testTime(),
-			url:       "/10583?",
-			res: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				return spec
-			}(),
-		},
-	}
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprint(w, `{"AnnotatorDate": "2018-12-05T00:00:00Z", "Annotations": {"127.0.0.1" : {"Geo":{"continent_code":"","country_code":"US","country_code3":"USA","country_name":"United States of America","region":"NY","metro_code":0,"city":"Scarsdale","area_code":10583,"postal_code":"10583","latitude":41.0051,"longitude":73.7846},"Network":{}}`+
-			`,"1.0.0.127" : {"Geo":{"continent_code":"","country_code":"US","country_code3":"USA","country_name":"United States of America","region":"NY","metro_code":0,"city":"Scarsdale","area_code":10584,"postal_code":"10584","latitude":41.0051,"longitude":73.7846},"Network":{}}}}`)
-	}))
-	defer ts.Close()
-	for _, test := range tests {
-		annotation.BatchURL = ts.URL + test.url
-		p.AddGeoDataNDTConnSpec(test.spec, test.timestamp)
-		if diff := deep.Equal(test.spec, test.res); diff != nil {
-			log.Println(test.spec)
-			t.Error(test.spec, diff)
-		}
-	}
-}
-
-func TestCopyStructToMap(t *testing.T) {
-	tests := []struct {
-		source interface{}
-		dest   map[string]bigquery.Value
-		res    map[string]bigquery.Value
-	}{
-		{
-			source: &struct {
-				A   int64
-				Bee string
-			}{A: 1, Bee: "2"},
-			dest: make(map[string]bigquery.Value),
-			res:  map[string]bigquery.Value{"a": int64(1), "bee": "2"},
-		},
-		{
-			source: &struct {
-				A   int64
-				Bee string
-			}{A: 0, Bee: ""},
-			dest: make(map[string]bigquery.Value),
-			res:  map[string]bigquery.Value{},
-		},
-		{
-			source: &struct{}{},
-			dest:   make(map[string]bigquery.Value),
-			res:    map[string]bigquery.Value{},
-		},
-	}
-	for _, test := range tests {
-		p.CopyStructToMap(test.source, test.dest)
-		if diff := deep.Equal(test.dest, test.res); diff != nil {
-			t.Error(diff)
-		}
-	}
-
-}
-
-func TestGetAndInsertTwoSidedGeoIntoNDTConnSpec(t *testing.T) {
-	tst, _ := time.Parse(time.RFC3339, "2002-10-02T15:00:00Z")
-	tests := []struct {
-		spec      schema.Web100ValueMap
-		timestamp time.Time
-		url       string
-		res       schema.Web100ValueMap
-	}{
-		{
-			spec: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				spec["client_ip"] = "127.0.0.1"
-				spec["server_ip"] = "1.0.0.127"
-				return spec
-			}(),
-			timestamp: tst,
-			url:       "/10583?",
-			res: func() schema.Web100ValueMap {
-				spec := schema.EmptyConnectionSpec()
-				spec["client_ip"] = "127.0.0.1"
-				spec["server_ip"] = "1.0.0.127"
-				geoc := spec.Get("client_geolocation")
-				geoc["country_code"] = "US"
-				geoc["country_code3"] = "USA"
-				geoc["country_name"] = "United States of America"
-				geoc["region"] = "NY"
-				geoc["city"] = "Scarsdale"
-				geoc["area_code"] = int64(10583)
-				geoc["postal_code"] = "10583"
-				geoc["latitude"] = float64(41.0051)
-				geoc["longitude"] = float64(73.7846)
-				geos := spec.Get("server_geolocation")
-				geos["country_code"] = "US"
-				geos["country_code3"] = "USA"
-				geos["country_name"] = "United States of America"
-				geos["region"] = "NY"
-				geos["city"] = "Scarsdale"
-				geos["area_code"] = int64(10584)
-				geos["postal_code"] = "10584"
-				geos["latitude"] = float64(41.0051)
-				geos["longitude"] = float64(73.7846)
-
-				cn := spec.Get("client").Get("network")
-				cn["asn"] = int64(123)
-				sn := spec.Get("server").Get("network")
-				sn["asn"] = int64(456)
-
-				return spec
-			}(),
-		},
-	}
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprint(w, `{"AnnotatorDate": "2018-12-05T00:00:00Z", "Annotations": {"127.0.0.1" : {"Geo":{"continent_code":"","country_code":"US","country_code3":"USA","country_name":"United States of America","region":"NY","metro_code":0,"city":"Scarsdale","area_code":10583,"postal_code":"10583","latitude":41.0051,"longitude":73.7846},"Network":{"Systems":[{"ASNs":[123]}]}}`+
-			`,"1.0.0.127" : {"Geo":{"continent_code":"","country_code":"US","country_code3":"USA","country_name":"United States of America","region":"NY","metro_code":0,"city":"Scarsdale","area_code":10584,"postal_code":"10584","latitude":41.0051,"longitude":73.7846},"Network":{"Systems":[{"ASNs":[456]}]}}}}`)
-	}))
-	defer ts.Close()
-	for _, test := range tests {
-		annotation.BatchURL = ts.URL + test.url
-		p.GetAndInsertTwoSidedGeoIntoNDTConnSpec(test.spec, test.timestamp)
-		if diff := deep.Equal(test.spec, test.res); diff != nil {
-			t.Error(test.url, diff)
 		}
 	}
 }
